game compiles. must still adapt overla
